SI-9208 Phase out ad-hoc, incorrect implicit shadowing detection
Paraphrasing Greenspun:

> Any sufficiently complicated Implicit Search program contains an
> ad hoc, informally-specified, bug-ridden, slow implementation of
> half of `typedIdent`.

The shadowing detection code that is being phased out was exactly
this. We have since heeded the aphorism above and perform
shadowing detection after typechecking a candidate implicit
(in 9129cfe911).

But, we didn't remove the pre-filtering of candidated based on
the old shadowing checks. This commit does exactly that. As this
introduces a minor source incompatibility, it is predicated on
-Xsource:2.12.

To effect this change, I also had to remedy a latent bug in the code
that gathers in-scope implicitss. Implicit parameters of a class
were being reported twice with the mistake filtered out by the
shadowing filter.