Improve diagnostic error on failed genLoadIdent
This error has been haunting us recently, firstly on Greg's machine
when compiling the compiler using the new SBT build, and more recently
during PR validation in #4316.

This commit will output an error like:

  symbol value c#16812 does not exist in Macro.impl, which contains locals value a#16810, value b#16811

I've included symbol IDs in the hope that they will prove useful.

It seems that synthetic identifiers generated by the pattern matcher
are often seen in the vicinity of this bug.