Clean up the way compiler settings are accessed in the backend.
Many backend components don't have access to the compiler instance
holding the settings.

Before this change, individual settings required in these parts of the
backend were made available as members of trait BTypes, implemented
in the subclass BTypesFromSymbols (which has access to global).

This change exposes a single value of type ScalaSettings in the super
trait BTypes, which gives access to all compiler settings.