Don't inlinie if the resulting method becomes too large for the JVM
This threshold is really the last resort and should never be reached.
An inlining heuristic that blows up methods to the maximum size
allowed by the JVM is broken.

In the future we need to include some heuristic about code size when
making an inlining decision, see github.com/scala-opt/scala/issues/2