fix scaladoc issue with parsing of empty tags
Consider the following code:

/**
  * @see
  * @deprecated
  */
  object Foo

The comment parser properly parsed the body of the 'see' tag as
empty, but not the one of 'deprecated': it supposedly contains a single
character, a newline '\n', which is wrong.

This always happened to the last tag in the list; it is always appended
a new line (whether empty or not), which breaks formatting (and things
later on that test if a body is empty of not).