SI-9267: Constructor properties are not properly copied
The parameterNamed(name: Name) method in Constructors currently
selects the first result from a list of matches, which allows
for an invalid assignment.

A fix proposal is to use an exact match instead of the startsWith
method, while also adding a sanity check in case of multiple
matched constructor parameters.