SI-9264 combat intermittent disappearing valdefs
Since at least 2.11.6, we seem to sometimes drop
ValDefs that are actually needed. One theory is that
this method was somehow dropping vals that shouldn't be??

Let's see if this makes the spurious disappearances go away.

This adds about 2% to bytecode size, but if the compiler goes
back to is old predictable ways, it's worth it.

TODO: what's the underlying cause? see ticker for more speculation