SI-9273 Avoid unpositioned error for bare classOf
A bare identifier `classOf` in a position wth an expected type of
`Class[_]` was leading to an unpositioned error.

This is due to special treatment of bare `classOf` in `typedIdent`
creating an ephemeral, but unpositioned, `TypeTree`.

This commit positions that tree and tests that the error is
issued at a sensible position.

There is still an irregularity between `classOf` and `Predef.classOf`,
but that seems esoteric enough to leave alone for now.