SI-9268 Be robust against absent classfiles during signature parsing
When parsing a Java generic signature that references an inner
class `A$B`, we were tripping an assertion if the enclosing class
`A` was absent.

This commit creates a stub symbol for `B` when this happens, rather
than continuing on with `NoSymbol`.

The enclosed test shows that we can instantiate a class containing
a method referring to such an inner class with only a warning
about the absent classfile, and that an error is issued only upon
a subsequent attempt to call the method.