SI-9298 Fix erasure of value classes in Java
Value classes that appear in signatures of Java defined methods
should not be erased to the underlying type.

Before this change, we'd get a `ClassCastException`, as the Scala
call site would unbox the value class despite the fact the Java
recipient would expect the boxed representation.

I've tested this for primitive and object wrapped types in parameter
and return position.