Avoid inefficient specialied lambdas w. delambdafy jvm-1.8, GenASM
A previous change disabled -Ydelambdafy:method for specialized
lambdas, as `DelambdafyTransformer` made no attempt to emit
the requisite machinery to avoid boxing.

This was loosened to allow them under `-target:jvm-1.8`, in the
knowledge that `indylambda` would do the right thing.

However, this wasn't quite right: indylambda is only supported
in `GenBCode`, so we should consider that setting as well.