SI-7773 Restore phase id to icode filename
Having icode output files of the form `X-24.icode` went in and
out of style using the long-form phase name because it broke
the windows nightly build somehow. Here's hoping using just the
phase id works on this year's infrastructure.

As previously, the long name is still available under `-Ydebug`,
because why not debug.