Use the scala-jline fork
The scala compiler depending on a jline release pollutes the classpath
of projects that embed the scala compiler or REPL, such as the spark
REPL. The scala-jline fork is identical to official jline, but with
the package name changed to scala.tools.jline.

For reference, the PR that removed jline from the scala repository
and introduced stock jline was #2704.