Centralize dependencies on jline
Code that depends on jline is now in package `scala.tools.nsc.interpreter.jline`.

To make this possible, remove the `entries` functionality from `History`,
and add the `historicize` method. Also provide an overload for `asStrings`.

Clean up a little along the way in `JLineHistory.scala` and `JLineReader.scala`.

Next step: fall back to an embedded jline when the expected jline jar
is not on the classpath.

The gist of the refactor: https://gist.github.com/adriaanm/02e110d4da0a585480c1