SI-1931 Hide Predef.any2stringadd in REPL
User imports that reference Predef are relocated to the top of
the wrapping template so that they can hide implicits defined
in Predef.

Only one import from Predef is retained for special treatment.
This is simple and sane. The test shows that `import Predef._`
restores Predef implicits even if a user-defined term would
normally be in scope.

A smart `:import` command to turn off or quarantine imports explicitly
would allow fine-grained control.