SI-9206 BooleanProp if set and not untrue
Previously, handy `sys.BooleanProp.keyExists` ignored the
property value. While trying not to make any real estate puns,
this commit will let it go false if a value is supplied that
is not true in the usual Java sense. But what is truth?

Allows `scala -Dscala.color=off`, for example.