SI-9359 Fix InnerClass entry flags for nested Java enums
The access flags in InnerClass entries for nested Java enums were
basically completely off.

A first step is to use the recently introduced backend method
`javaClassfileFlags`, which is now moved to BCodeAsmCommon.
See its doc for an explanation.

Then the flags of the enum class symbol were off. An enum is
  - final if none of its values has a class body
  - abstract if it has an abstract method
(https://docs.oracle.com/javase/specs/jls/se7/html/jls-8.html#jls-8.9)

When using the ClassfileParser:
  - ENUM was never added. I guess that's just an oversight.
  - ABSTRACT (together with SEALED) was always added. This is to
    enable exhaustiveness checking, see 3f7b8b5. This is a hack and we
    have to go through the class members in the backend to find out if
    the enum actually has the `ACC_ABSTRACT` flag or not.

When using the JavaParser:
  - FINAL was never added.
  - ABSTRACT was never added.

This commit fixes all of the above and tests cases (Java enum read
from the classfile and from source).