SI-8892 Fix incorrect qualification in error message
Since beadafa2, we've ended up with nutty error messages for
type errors involving aliases that target types in `java.lang` and
`scala` packages. I think the intent of that change was to force
the error messages to qualify types like `String` when needed, but
to leave them unqualified by default.

However, this led to this flat out wrong message in the enclosed test.

	 found   : B
	 required: C.this.java.lang.B
	    (which expands to)  String

I've changed the heuristic slightly limit this code to aliases
that are eponymous with their targets. Still feels pretty kludgy, but
we can at least book a little progress.