Cast arguments where necessary before rewriting closure invocations
The parameter types of a closure invocation can be supertypes of the
parameter types in the implementation method. The closure automatically
casts arguments to the right type. We need to do the same when
rewriting closure invocations. Example:

    val fun: String => String = l => l
    val l = List("")
    fun(l.head)

The closure object calls `apply(Object)Object`. The body method takes
an argument of type `String`.