[sbt] Add a project for repl-jline
Followup to the recent restructure in the Ant build. This commit
adds a project for repl-jline, and packages its output in the
JAR created by the repl project.

As per the Ant build, we also use JarJar to embed a shaded copy
of jline (with a compatible copy of our jline facade).

During the process, I've also:

  - Updated the the latest SBT release
  - Disabled `repl/run`, as it does not have JLine
  - Copied `jline.jar` to `pack/lib` so the distro-like version
    of the REPL created by `dist/mkBin` has completion.
  - Added forkjoin classes to the scala-library.jar
  - Improved diagnostic logging in REPL startup to print the full
    exception stack traces, causes and all.