Improved message for missing argument list
Clarifies the language and rules for eta-expansion.

A missing argument in a list, as opposed to a missing
argument list, results in a different message.

The comical expansion in parens does not attempt to show
what was already applied, but succeeds in showing at a
glance the shape of the method in question.

```
scala> def m(i: Int, j: Int)(x: Int) = ???
m: (i: Int, j: Int)(x: Int)Nothing

scala> m
<console>:12: error: missing argument list for method m
Unapplied methods are only converted to functions when a function type is expected.
You can make this conversion explicit by writing `m _` or `m(_,_)(_)` instead of `m`.
       m
       ^
```

The original submission was due to sschaef and the wording due
to adriaanm, with a minor tweak.