GenBCode: fix incrementatal compilation by mimicing GenASM
The incremental compiler in SBT uses an implementation detail
of the compiler backend to enumerate the classes that are actually
written to disk.

This commit mimics this in GenBCode by populating `Run#icode` with
an `IClass` for each `ClassDef` processed.

We should revisit this by creating a dedicated API for this purpose
and migrating SBT to use that. We should also revisit this code
as we implement closure elimination in the GenBCode; this commit
assumes that all `ClassDef`s that enter the backend will generate
classfile products.

The enclosed test is extracted from the incrementatl compiler.

I've also manually integration tested this with SBT:

  https://gist.github.com/retronym/fabf6f92787ea9c1ce67