SI-6810 Disallow EOL in char literal
It's clear that char literals are one-lined like normal
string literals.

By the same token, pun intended, char literals accept
unicode escapes the same as string literals, including
`\u000A`.

This commit adds the usual exclusions (CR, NL, SU).

The spec is outdated in outlawing chars that are not
"printable", in particular, the ASCII control codes.
The original intention may have been that the ordinary
string escapes are required, such as "\b\n". Note that
some common escapes are absent, such as "\a".